+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - **Fix**: Copied files are now immediately indexed
 - **Fix**: Improved and 20x faster planet database setup
+- **Feature**: Huawei Moving Picture support
 
 ## [v8.0.0] - 2026-04-04